made 4 different components for the editprofile section
