+++ src/profile/EditProfile.js
@@ -1,11 +1,9 @@
 import React, { Component } from "react";
-import FormGroup from "react-bootstrap/lib/FormGroup";
-import ControlLabel from "react-bootstrap/lib/ControlLabel";
-import FormControl from "react-bootstrap/lib/FormControl";
+import AboutYou from "/home/antonio/paretOS/src/profile/AboutYou";// need to change path
+import Project from "/home/antonio/paretOS/src/profile/Project";// need to change path
+import ProfileId from "/home/antonio/paretOS/src/profile/ProfileId";// need to change path
+import Languages from "/home/antonio/paretOS/src/profile/Languages";// need to change path
 import Button from "react-bootstrap/lib/Button";
-import Glyphicon from "react-bootstrap/lib/Glyphicon";
-import Image from "react-bootstrap/lib/Image";
-import LoaderButton from "../components/LoaderButton";
 import uuidv4 from "uuid";
 import API from "@aws-amplify/api";
 import { I18n } from "@aws-amplify/core";
@@ -13,8 +11,8 @@ import Storage from "@aws-amplify/storage";
 import generator from "generate-password";
 import { errorToast } from "../libs/toasts";
 import Tour from "reactour";
-import question from "../assets/help.png";
 import "react-quill/dist/quill.snow.css";
+
 
 /**
  * These are the forms where you can edit your profile.
@@ -27,9 +25,7 @@ export default class EditProfile extends Component {
     super(props);
 
     this.state = {
-      isLoading: false,
       users: [],
-      providers: [],
       summary: "",
       summaryCheck: false,
       user: {
@@ -63,38 +59,6 @@ export default class EditProfile extends Component {
         "https://wallsheaven.co.uk/photos/A065336811/220/user-account-profile-circle-flat-icon-for-apps-and-websites-.webp",
     };
   }
-
-  async componentDidMount() {
-    const path = window.location.pathname.split("/");
-    let userId;
-    if (window.location.pathname === "/") {
-      userId = this.props.username;
-      this.setState({
-        user: this.props.user,
-        id: userId,
-        summary: this.props.user.summary,
-        fName: this.props.user.fName,
-        lName: this.props.user.lName,
-        note: this.props.user.notes[0],
-        school: this.props.user.school,
-        defaultLanguage: this.props.user.defaultLanguage,
-      });
-    } else {
-      userId = path[path.length - 1];
-      const response = await API.get("pareto", `/users/${userId}`);
-      this.setState({
-        user: response[0],
-        id: userId,
-        summary: response[0].summary,
-        fName: response[0].fName,
-        lName: response[0].lName,
-        note: response[0].notes[0],
-        school: response[0].school,
-        defaultLanguage: response[0].defaultLanguage,
-      });
-    }
-  }
-
   closeTour = () => {
     this.setState({
       isTourOpen: false,
@@ -105,26 +69,6 @@ export default class EditProfile extends Component {
     this.setState({
       [event.target.id]: event.target.value,
     });
-  };
-
-  handleSubmit = async (event) => {
-    event.preventDefault();
-
-    this.setState({ isLoading: true });
-
-    let body = {
-      summary: this.state.summary,
-    };
-    try {
-      const response = await API.put("pareto", `/users/${this.state.id}`, {
-        body,
-      });
-      this.setState({ user: response, summaryCheck: false });
-    } catch (e) {
-      console.log("email send error: ", e);
-      // toast(e, { type: toast.TYPE.ERROR})
-    }
-    this.setState({ isLoading: false });
   };
 
   editName = async () => {
@@ -142,36 +86,20 @@ export default class EditProfile extends Component {
     }
   };
 
-  updateSchool = async () => {
-    let body = {
-      school: this.state.school,
-    };
-    try {
-      const response = await API.put("pareto", `/users/${this.state.id}`, {
-        body,
-      });
-      console.log(response);
-      this.setState({ user: response });
-    } catch (e) {
-      console.log(e);
-    }
-  };
-
-  updateLanguage = async () => {
-    let body = {
-      defaultLanguage: this.state.defaultLanguage,
-    };
-    try {
-      const response = await API.put("pareto", `/users/${this.state.id}`, {
-        body,
-      });
-      console.log(response);
-      this.setState({ user: response });
-    } catch (e) {
-      console.log(e);
-    }
-  };
-
+  // updateSchool = async () => {
+  //   let body = {
+  //     school: this.state.school,
+  //   };
+  //   try {
+  //     const response = await API.put("pareto", `/users/${this.state.id}`, {
+  //       body,
+  //     });
+  //     console.log(response);
+  //     this.setState({ user: response });
+  //   } catch (e) {
+  //     console.log(e);
+  //   }
+  // };
   addProject = async () => {
     let projects = this.state.user.projects.slice();
 
@@ -206,40 +134,40 @@ export default class EditProfile extends Component {
     }
   };
 
-  onChange = async (e) => {
-    const file = e.target.files[0];
-    let fileType = e.target.files[0].name.split(".");
+  // onChange = async (e) => {
+  //   const file = e.target.files[0];
+  //   let fileType = e.target.files[0].name.split(".");
 
-    try {
-      // @TODO: check to see whether this works for video, and what safeguards may not to be added.
-      // @TODO: update this manual id for a dynamically generated one
-      let pictureKey = await Storage.put(
-        `48150f13-679a-40fc-9f23-b4ebe4af6f08.${fileType[1]}`,
-        file,
-        {
-          contentType: "image/*",
-          bucket: process.env.REACT_APP_PHOTO_BUCKET,
-        }
-      );
-      console.log("Key: ", pictureKey);
-      let updatedProfile = await API.put(
-        "pareto",
-        `/users/${this.state.user.id}`,
-        {
-          body: {
-            picture: `https://${process.env.REACT_APP_PHOTO_BUCKET}.s3.amazonaws.com/public/${pictureKey.key}`,
-          },
-        }
-      );
-      console.log(updatedProfile);
-      this.setState({
-        user: updatedProfile,
-      });
-      // need to save the key
-    } catch (e) {
-      errorToast(e);
-    }
-  };
+  //   try {
+  //     // @TODO: check to see whether this works for video, and what safeguards may not to be added.
+  //     // @TODO: update this manual id for a dynamically generated one
+  //     let pictureKey = await Storage.put(
+  //       `48150f13-679a-40fc-9f23-b4ebe4af6f08.${fileType[1]}`,
+  //       file,
+  //       {
+  //         contentType: "image/*",
+  //         bucket: process.env.REACT_APP_PHOTO_BUCKET,
+  //       }
+  //     );
+  //     console.log("Key: ", pictureKey);
+  //     let updatedProfile = await API.put(
+  //       "pareto",
+  //       `/users/${this.state.user.id}`,
+  //       {
+  //         body: {
+  //           picture: `https://${process.env.REACT_APP_PHOTO_BUCKET}.s3.amazonaws.com/public/${pictureKey.key}`,
+  //         },
+  //       }
+  //     );
+  //     console.log(updatedProfile);
+  //     this.setState({
+  //       user: updatedProfile,
+  //     });
+  //     // need to save the key
+  //   } catch (e) {
+  //     errorToast(e);
+  //   }
+  // };
 
   render() {
     const steps = [
@@ -254,81 +182,9 @@ export default class EditProfile extends Component {
     ];
     return (
       <div className="flex-down">
-        <div className="flex">
-          <div className="first-step-home">
-            {this.state.editName === false ? (
-              <div className="flex">
-                <img
-                  src={this.state.user.picture || this.state.picture}
-                  height="50"
-                  width="50"
-                  alt="Profile"
-                  style={{ marginTop: 26 }}
-                />
-
-                <h1>{this.state.user.fName}</h1>
-                <Glyphicon
-                  onClick={() => this.setState({ editName: true })}
-                  glyph="glyphicon glyphicon-pencil"
-                  height="33"
-                  width="33"
-                  style={{ marginTop: 54, marginLeft: 6, cursor: "pointer" }}
-                />
-                <Image
-                  src={question}
-                  onClick={(event) => {
-                    event.preventDefault();
-                    this.setState({ isTourOpen: true });
-                  }}
-                  height="50"
-                  width="50"
-                  circle
-                  style={{
-                    cursor: "pointer",
-                    marginTop: 30,
-                    marginLeft: 40,
-                  }}
-                />
-              </div>
-            ) : (
-              <div className="flex">
-                <div className="flex-down" style={{ marginTop: 20 }}>
-                  <div className="flex">
-                    <FormGroup controlId="fName" bsSize="large">
-                      <ControlLabel>{I18n.get("firstName")}</ControlLabel>
-                      <FormControl
-                        value={this.state.fName}
-                        onChange={this.handleChange}
-                      />
-                    </FormGroup>
-                    <FormGroup controlId="lName" bsSize="large">
-                      <ControlLabel>{I18n.get("lastName")}</ControlLabel>
-                      <FormControl
-                        value={this.state.lName}
-                        onChange={this.handleChange}
-                      />
-                    </FormGroup>
-                  </div>
-                  <Button onClick={this.editName}>
-                    {I18n.get("editName")}
-                  </Button>
-                </div>
-                <div className="flex-down" style={{ marginTop: 28 }}>
-                  <h3>{I18n.get("changePicture")}</h3>
-                  <input
-                    type="file"
-                    accept="image/*"
-                    onChange={(evt) => this.onChange(evt)}
-                  />
-                  <Button onClick={() => this.setState({ editName: false })}>
-                    {I18n.get("cancel")}
-                  </Button>
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-
+    
+      <ProfileId />
+      
         {/* Hiding the below for now, not sure it makes sense for users to update. More so for admin? Is this feature even relevant anymore? could be for organizations, but not on a user-level */}
 
         {/* {this.state.editSchool === false ? (
@@ -370,149 +226,16 @@ export default class EditProfile extends Component {
           </div>
         )} */}
         <div>
-          <h2>About you</h2>
-          {this.state.summaryCheck ? (
-            <React.Fragment>
-              <FormGroup controlId="summary" bsSize="large">
-                <ControlLabel>{I18n.get("bio")}</ControlLabel>
-                <FormControl
-                  value={this.state.summary}
-                  onChange={this.handleChange}
-                  componentClass="textarea"
-                />
-              </FormGroup>
-              <div className="flex">
-                <Button onClick={() => this.setState({ summaryCheck: false })}>
-                  {I18n.get("cancel")}
-                </Button>
-
-                <LoaderButton
-                  align="center"
-                  block
-                  bsSize="small"
-                  type="submit"
-                  // disabled={!this.validateForm()}
-                  onClick={this.handleSubmit}
-                  isLoading={this.state.isLoading}
-                  text="Update Summary"
-                  loadingText="Creation"
-                />
-              </div>
-            </React.Fragment>
-          ) : (
-            <React.Fragment>
-              <div className="block">
-                <p>
-                  {this.state.user.summary}{" "}
-                  <Glyphicon
-                    onClick={() => this.setState({ summaryCheck: true })}
-                    glyph="glyphicon glyphicon-pencil"
-                    height="33"
-                    width="33"
-                    style={{ marginLeft: 6, cursor: "pointer" }}
-                  />
-                </p>
-              </div>
-            </React.Fragment>
-          )}
+          <AboutYou />
         </div>
-
         <div>
-          <h2 className="third-step-home">
-            {I18n.get("projects")}{" "}
-            <Glyphicon
-              onClick={() =>
-                this.setState({ addProject: !this.state.addProject })
-              }
-              glyph="glyphicon glyphicon-plus"
-              height="33"
-              width="33"
-              style={{ marginLeft: 4, cursor: "pointer", marginTop: 2 }}
-            />
-          </h2>
-          {this.state.user.projects.length < 1 ? (
-            <p className="block">{I18n.get("noProjectsYet")}</p>
-          ) : (
-            <React.Fragment>
-              {this.state.user.projects.map((project, i) => {
-                return (
-                  <div className="block">
-                    <h3>{project.name}</h3>
-                    <p>{project.description}</p>
-                    <a href={project.github} target="_blank" rel="noopener">
-                      GitHub Link Here
-                    </a>
-                  </div>
-                );
-              })}
-            </React.Fragment>
-          )}
-          {this.state.addProject ? (
-            <div className="block">
-              <FormGroup controlId="name" bsSize="large">
-                <ControlLabel>{I18n.get("projectName")}</ControlLabel>
-                <FormControl
-                  value={this.state.name}
-                  onChange={this.handleChange}
-                />
-              </FormGroup>
-
-              <FormGroup controlId="description" bsSize="large">
-                <ControlLabel>{I18n.get("description")}</ControlLabel>
-                <FormControl
-                  value={this.state.description}
-                  onChange={this.handleChange}
-                />
-              </FormGroup>
-              <FormGroup controlId="github" bsSize="large">
-                <ControlLabel>{I18n.get("githubRepository")}</ControlLabel>
-                <FormControl
-                  value={this.state.github}
-                  onChange={this.handleChange}
-                />
-              </FormGroup>
-              <div className="flex">
-                <Button onClick={() => this.setState({ addProject: false })}>
-                  {I18n.get("cancel")}
-                </Button>
-
-                <Button onClick={this.addProject}>{I18n.get("save")}</Button>
-              </div>
-            </div>
-          ) : null}
+          <Project />
         </div>
-
-        <br />
-
-        <FormGroup controlId="defaultLanguage" bsSize="large">
-          <ControlLabel>Default Language</ControlLabel>
-          <div className="flex">
-            <FormControl
-              componentClass="select"
-              onChange={this.handleChange}
-              value={this.state.defaultLanguage}
-            >
-              <option value="en">Choose Here</option>
-              <option value="en">English</option>
-              <option value="lg">Luganda</option>
-              <option value="ac">Acholi</option>
-              <option value="es">Spanish</option>
-            </FormControl>
-            <LoaderButton
-              align="center"
-              block
-              type="submit"
-              style={{ width: 90 }}
-              // disabled={!this.validateForm()}
-              onClick={this.updateLanguage}
-              isLoading={this.state.isLoading}
-              text="Save"
-              loadingText="Updating..."
-            />
-          </div>
-        </FormGroup>
-
-        <Button onClick={() => this.props.history.push("/settings/password")}>
+          <br />
+        <div>
+          <Languages />
+        </div>
+       <Button onClick={() => this.props.history.push("/settings/password")}>
           Change Password
         </Button>
